+++ gh-pages/component/MilkdownEditor/MilkdownEditor.tsx
@@ -1,24 +1,35 @@
 import React from 'react';
+
 import { defaultValueCtx, Editor, editorViewOptionsCtx, rootCtx } from '@milkdown/core';
-import { ReactEditor, useEditor } from '@milkdown/react';
-import { listener, listenerCtx } from '@milkdown/plugin-listener';
-import { gfm } from '@milkdown/preset-gfm';
-import { history } from '@milkdown/plugin-history';
+
 import { clipboard } from '@milkdown/plugin-clipboard';
+import { history } from '@milkdown/plugin-history';
+import { listener, listenerCtx } from '@milkdown/plugin-listener';
 import { prism } from '@milkdown/plugin-prism';
-import { tooltip } from '@milkdown/plugin-tooltip';
-import { math } from '@milkdown/plugin-math';
-import { slash } from '@milkdown/plugin-slash';
+
+import { emoji } from '@milkdown/plugin-emoji';
+import '@milkdown/plugin-emoji/lib/style.css';
+
 import { cursor } from '@milkdown/plugin-cursor';
-import className from './style.module.css';
+import '@milkdown/plugin-cursor/lib/style.css';
+
+import { math } from '@milkdown/plugin-math';
+import '@milkdown/plugin-math/lib/style.css';
+
+import { slash } from '@milkdown/plugin-slash';
+import '@milkdown/plugin-slash/lib/style.css';
+
+import { tooltip } from '@milkdown/plugin-tooltip';
+import '@milkdown/plugin-tooltip/lib/style.css';
+
+import { gfm } from '@milkdown/preset-gfm';
+import '@milkdown/preset-gfm/lib/style.css';
+
+import { ReactEditor, useEditor } from '@milkdown/react';
 
 import '@milkdown/theme-nord/lib/theme.css';
-import '@milkdown/preset-gfm/lib/style.css';
-import '@milkdown/plugin-math/lib/style.css';
-import '@milkdown/plugin-table/lib/style.css';
-import '@milkdown/plugin-tooltip/lib/style.css';
-import '@milkdown/plugin-slash/lib/style.css';
-import '@milkdown/plugin-cursor/lib/style.css';
+
+import className from './style.module.css';
 
 type Props = {
     content: string;
@@ -42,6 +53,7 @@ export const MilkdownEditor: React.FC<Props> = ({ content, readOnly, onChange })
                 .use(history)
                 .use(cursor())
                 .use(prism)
+                .use(emoji)
                 .use(tooltip)
                 .use(math);