+++ packages/vue/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @milkdown/vue
 
+## 4.8.1
+
+### Patch Changes
+
+-   6d801db: Refactor for small improvements.
+-   Updated dependencies [6d801db]
+    -   @milkdown/utils@4.8.1
+
+## 4.8.0
+
+### Minor Changes
+
+-   501ad84: Add emoji plugin
+
+### Patch Changes
+
+-   Updated dependencies [501ad84]
+    -   @milkdown/utils@4.8.0
+
 ## 4.7.1
 
 ### Patch Changes